[tests/e2e] refactor into sub-mods
