+++ tests/src/lib.rs
@@ -3,6 +3,7 @@
 
 mod vm_host_env;
 pub use vm_host_env::{tx, vp};
+#[cfg(test)]
 mod e2e;
 
 /// Using this import requires `tracing` and `tracing-subscriber` dependencies.
